Fix size_t/unsigned warnings: use (unsigned) for short string's length.

+++ lockfile.c
@@ -70,7 +70,7 @@ lockfile_acquire(const char *path, unsigned staleness_limit)
 				}
 			}
 			free(content);
-			const size_t bufsize = 1024;
+			const unsigned int bufsize = 1024;
 			content = x_malloc(bufsize);
 			int len = read(fd, content, bufsize - 1);
 			if (len == -1) {
@@ -82,7 +82,7 @@ lockfile_acquire(const char *path, unsigned staleness_limit)
 			content[len] = '\0';
 		} else {
 			// We got the lock.
-			if (write(fd, my_content, strlen(my_content)) == -1) {
+			if (write(fd, my_content, (unsigned)strlen(my_content)) == -1) {
 				cc_log("lockfile_acquire: write %s: %s", lockfile, strerror(errno));
 				close(fd);
 				x_unlink(lockfile);
